bot_service/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==25.1.0
    # via aiogram
aiogram==3.27.0
    # via bot-service (pyproject.toml)
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.5
    # via aiogram
aiosignal==1.4.0
    # via aiohttp
amqp==5.3.1
    # via kombu
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==26.1.0
    # via aiohttp
billiard==4.2.4
    # via celery
celery==5.6.3
    # via bot-service (pyproject.toml)
certifi==2026.2.25
    # via
    #   aiogram
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   uvicorn
cryptography==46.0.7
    # via python-jose
ecdsa==0.19.2
    # via python-jose
fakeredis==2.35.1
    # via bot-service (pyproject.toml)
fastapi==0.136.0
    # via bot-service (pyproject.toml)
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.4.0
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   bot-service (pyproject.toml)
    #   respx
idna==3.11
    # via
    #   anyio
    #   httpx
    #   yarl
iniconfig==2.3.0
    # via pytest
kombu==5.6.2
    # via celery
magic-filter==1.0.12
    # via aiogram
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
packaging==26.1
    # via
    #   kombu
    #   pytest
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via click-repl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.3
    # via
    #   python-jose
    #   rsa
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   aiogram
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via bot-service (pyproject.toml)
pygments==2.20.0
    # via pytest
pytest==9.0.3
    # via
    #   bot-service (pyproject.toml)
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==1.3.0
    # via bot-service (pyproject.toml)
pytest-mock==3.15.1
    # via bot-service (pyproject.toml)
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.5.0
    # via bot-service (pyproject.toml)
pyyaml==6.0.3
    # via uvicorn
redis==7.4.0
    # via
    #   bot-service (pyproject.toml)
    #   fakeredis
respx==0.23.1
    # via bot-service (pyproject.toml)
rsa==4.9.1
    # via python-jose
ruff==0.15.11
    # via bot-service (pyproject.toml)
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sortedcontainers==2.4.0
    # via fakeredis
sqlalchemy==2.0.49
    # via bot-service (pyproject.toml)
starlette==1.0.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   aiogram
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
tzdata==2026.1
    # via
    #   kombu
    #   tzlocal
tzlocal==5.3.1
    # via celery
uvicorn==0.44.0
    # via bot-service (pyproject.toml)
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.6.0
    # via prompt-toolkit
websockets==16.0
    # via uvicorn
yarl==1.23.0
    # via aiohttp
